general cleanup and reorganizing
