+++ src/components/FolderCreator.tsx
@@ -6,7 +6,7 @@ import {
   findAllDashboardsWithinCurrentFolderStruc,
   getCurrentFolder,
   getSurroundings,
-} from "../functions/folderFunctions.ts";
+} from "../functions";
 import styles from "../styles/FolderCreator.module.css";
 import { FolderCreatorProps } from "../types";
 import type { Folder, MenuObject } from "../types/index.ts";
@@ -65,7 +65,7 @@ export const FolderCreator = ({
     }
 
     setMenuObject((prevMenuObj) => {
-      const newMenuObj: MenuObject = JSON.parse(JSON.stringify(prevMenuObj));
+      const newMenuObj: MenuObject = structuredClone(prevMenuObj);
       const currentInd = newMenuObj?.currentFolder;
       if (newMenuObj.currentFolder.length === 0) {
         newMenuObj.folders[newFolderName] = {};
@@ -115,7 +115,7 @@ export const FolderCreator = ({
               className={styles["folder-creator-breadcrumb-link"]}
               onClick={() => {
                 setMenuObject((prevMenuObj) => {
-                  const newMenuObj = JSON.parse(JSON.stringify(prevMenuObj));
+                  const newMenuObj = structuredClone(prevMenuObj);
                   if (curr === "Home") {
                     newMenuObj.currentFolder = [];
                   } else {
@@ -169,7 +169,7 @@ export const FolderCreator = ({
         findAllDashboardsWithinCurrentFolderStruc(currentFolder);
 
       setMenuObject((prevMenuObj) => {
-        const newMenuObj: MenuObject = JSON.parse(JSON.stringify(prevMenuObj));
+        const newMenuObj: MenuObject = structuredClone(prevMenuObj);
         const currentInd = newMenuObj.currentFolder;
         if (currentInd.length === 0) {
           return newMenuObj;
@@ -254,7 +254,8 @@ export const FolderCreator = ({
               }`}
             >
               <p>
-                Type "DELETE" to delete this folder and subfolders. Dashboards will move up and out.
+                Type "DELETE" to delete this folder and subfolders. Dashboards
+                will move up and out.
               </p>
               <div className={styles["folder-delete-zone-confirm"]}>
                 <input
