+++ src/functions/formDiceResultString.ts
